requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
faiss-cpu
sentence-transformers
pypdf
python-dotenv
requests
langchain_google_genai
loguru